Add files via upload
In computer security and programming, a buffer overflow, or buffer overrun, is an anomaly where a program, while writing data to a buffer,
overruns the buffer's boundary and overwrites adjacent memory locations. This is a special case of the violation of memory safety.
Buffer overflows can be triggered by inputs that are designed to execute code or alter the way the program operates. This may result in erratic
program behavior, including memory access errors, incorrect results, a crash, or a breach of system security. Thus, they are the basis of many
software vulnerabilities and can be maliciously exploited.
Programming languages commonly associated with buffer overflows include C and C++, which provide no built-in protection against accessing or
overwriting data in any part of memory and do not automatically check that data written to an array (the built-in buffer type) is within the
boundaries of that array. Bounds checking can prevent buffer overflows